refactor: 운동 참여 상태  Enum 운동 도메인으로 변경
