+++ src/main/java/umc/cockple/demo/domain/exercise/domain/Guest.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import umc.cockple.demo.domain.exercise.dto.ExerciseGuestInviteDTO;
 import umc.cockple.demo.global.common.BaseEntity;
-import umc.cockple.demo.global.enums.ExerciseMemberShipStatus;
+import umc.cockple.demo.domain.exercise.enums.ExerciseMemberShipStatus;
 import umc.cockple.demo.global.enums.Gender;
 import umc.cockple.demo.global.enums.Level;
