+++ src/main/java/umc/cockple/demo/domain/exercise/enums/ExerciseMemberShipStatus.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.global.enums;
+package umc.cockple.demo.domain.exercise.enums;
 
 public enum ExerciseMemberShipStatus {
     PARTY_MEMBER,
